+++ src/apps/sequencer/engine/NoteTrackEngine.cpp
@@ -329,6 +329,11 @@ void NoteTrackEngine::monitorMidi(uint32_t tick, const MidiMessage &message) {
 
     if (_engine.recording() && _model.project().recordMode() == Types::RecordMode::StepRecord) {
         _stepRecorder.process(message, *_sequence, [this] (int midiNote) { return noteFromMidiNote(midiNote); });
+        if (Routing::isRouted(Routing::Target::CurrentRecordStep, _model.project().selectedTrackIndex())) {
+            _sequence->setCurrentRecordStep(_stepRecorder.stepIndex(), true);
+        } else {
+            _sequence->setCurrentRecordStep(_stepRecorder.stepIndex(), false);
+        }
     }
 }