+++ src/apps/sequencer/model/Routing.h
@@ -72,6 +72,7 @@ public:
         Divisor,
         Scale,
         RootNote,
+        CurrentRecordStep,
         Reseed,
         RestProbability2,
         RestProbability4,
@@ -121,6 +122,8 @@ public:
         case Target::Divisor:                   return "Divisor";
         case Target::Scale:                     return "Scale";
         case Target::RootNote:                  return "Root Note";
+        case Target::CurrentRecordStep:         return "REC Step";
+
         case Target::Reseed:                    return "Reseed";
         case Target::RestProbability2:          return "Rest Prob. 2";
         case Target::RestProbability4:          return "Rest Prob. 4";
@@ -182,7 +185,8 @@ public:
         case Target::RestProbability2:          return 34;
         case Target::RestProbability4:          return 35;
         case Target::RestProbability8:          return 36;
-        case Target::LengthModifier:            return 37;        
+        case Target::LengthModifier:            return 37;    
+        case Target::CurrentRecordStep:         return 38;    
 
         case Target::Last:                      break;
         }