+++ src/apps/sequencer/ui/pages/NoteSequenceEditPage.cpp
@@ -90,11 +90,14 @@ void NoteSequenceEditPage::draw(Canvas &canvas) {
     WindowPainter::drawActiveFunction(canvas, NoteSequence::layerName(layer()));
     WindowPainter::drawFooter(canvas, functionNames, pageKeyState(), activeFunctionKey());
 
-    const auto &trackEngine = _engine.selectedTrackEngine().as<NoteTrackEngine>();
+    auto &trackEngine = _engine.selectedTrackEngine().as<NoteTrackEngine>();
 
     auto &sequence = _project.selectedNoteSequence();
     const auto &scale = sequence.selectedScale(_project.scale());
     int currentStep = trackEngine.isActiveSequence(sequence) ? trackEngine.currentStep() : -1;
+    if (trackEngine.currentRecordStep()!=-1) {
+        trackEngine.setCurrentRecordStep(sequence.currentRecordStep());
+    }
     int currentRecordStep = trackEngine.isActiveSequence(sequence) ? trackEngine.currentRecordStep() : -1;
 
     const int stepWidth = Width / StepCount;
@@ -351,6 +354,8 @@ void NoteSequenceEditPage::keyPress(KeyPressEvent &event) {
     auto &sequence = _project.selectedNoteSequence();
     auto &track = _project.selectedTrack().noteTrack();
 
+    auto &trackEngine = _engine.selectedTrackEngine().as<NoteTrackEngine>();
+
     if (key.isContextMenu()) {
         contextShow();
         event.consume();
@@ -475,20 +480,37 @@ void NoteSequenceEditPage::keyPress(KeyPressEvent &event) {
 
     if (key.isLeft()) {
         if (key.shiftModifier()) {
-            _inMemorySequence = _project.selectedNoteSequence();
-            sequence.shiftSteps(_stepSelection.selected(), -1);
-            _stepSelection.shiftLeft();
+            if (trackEngine.currentRecordStep()!=-1) {
+                if (Routing::isRouted(Routing::Target::CurrentRecordStep, _model.project().selectedTrackIndex())) {
+                    sequence.setCurrentRecordStep(sequence.currentRecordStep()-1, true);
+                } else {
+                    sequence.setCurrentRecordStep(sequence.currentRecordStep()-1, false);
+                }
+            } else {
+                _inMemorySequence = _project.selectedNoteSequence();
+                sequence.shiftSteps(_stepSelection.selected(), -1);
+                _stepSelection.shiftLeft(sequence.lastStep()+1);
+            }
         } else {
-            track.setPatternFollowDisplay(false);
+             track.setPatternFollowDisplay(false);
              sequence.setSecion(std::max(0, sequence.section() - 1));
         }
         event.consume();
     }
     if (key.isRight()) {
         if (key.shiftModifier()) {
-            _inMemorySequence = _project.selectedNoteSequence();
-            sequence.shiftSteps(_stepSelection.selected(), 1);
-            _stepSelection.shiftRight();
+            if (trackEngine.currentRecordStep()!=-1) {
+                if (Routing::isRouted(Routing::Target::CurrentRecordStep, _model.project().selectedTrackIndex())) {
+                    sequence.setCurrentRecordStep(sequence.currentRecordStep()+1, true);
+                } else {
+                    sequence.setCurrentRecordStep(sequence.currentRecordStep()+1, false);
+                }
+                
+            } else {
+                _inMemorySequence = _project.selectedNoteSequence();
+                sequence.shiftSteps(_stepSelection.selected(), 1);
+                _stepSelection.shiftRight(sequence.lastStep()+1);
+            }
         } else {
             track.setPatternFollowDisplay(false);
             sequence.setSecion(std::min(3, sequence.section() + 1));
@@ -1060,7 +1082,6 @@ void NoteSequenceEditPage::tieNotes() {
                 showMessage("NOTES TIED");
             }
             sequence.step(i).setNote(sequence.step(first).note());
-            std::cerr << _stepSelection[i];
         }
     }
 }